% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.args.R
\name{r.args}
\alias{r.args}
\title{Write the arguments of a function selected}
\usage{
r.args()
}
\value{
arguments ready to run
}
\description{
Write the arguments of a function selected
}
\examples{

rnorm(n=100)


n=100;mean=1;sd=0


}
